requirements.txt
pydantic~=2.10
requests==2.31.0
beautifulsoup4==4.12.2
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
schedule==1.2.1
python-dotenv==1.0.0
tenacity==8.2.3
loguru==0.7.2
fastapi==0.104.1
uvicorn==0.24.0
httpx==0.25.1
python-dateutil
geopy>=2.0
